+++ src/main/kotlin/cc/loac/data/requests/MenuItemRequest.kt
@@ -7,7 +7,7 @@ import cc.loac.data.models.enums.MenuItemTarget
  */
 data class MenuItemRequest(
     /** 菜单项 ID */
-    val menuItemId: Int?,
+    val menuItemId: Long?,
     /** 菜单名称 */
     val displayName: String,
     /** 菜单地址 */
@@ -15,9 +15,9 @@ data class MenuItemRequest(
     /** 打开方式 */
     val target: MenuItemTarget? = MenuItemTarget.BLANK,
     /** 父菜单 ID */
-    val parentMenuId: Int,
+    val parentMenuId: Long,
     /** 父菜单项 ID */
-    val parentMenuItemId: Int?,
+    val parentMenuItemId: Long?,
     /** 菜单项排序索引 **/
     var index: Int
 )
