+++ src/main/kotlin/cc/loac/services/CategoryService.kt
@@ -18,7 +18,7 @@ interface CategoryService {
      * 删除分类
      * @param ids 分类 ID 集合
      */
-    suspend fun deleteCategories(ids: List<Int>): Boolean
+    suspend fun deleteCategories(ids: List<Long>): Boolean
 
     /**
      * 根据别名删除分类
@@ -48,7 +48,7 @@ interface CategoryService {
      * 根据分类 ID 获取分类
      * @param id 分类 ID
      */
-    suspend fun category(id: Int): Category?
+    suspend fun category(id: Long): Category?
 
     /**
      * 根据分类名称获取分类
